+++ src/azure-cli/azure/cli/command_modules/vm/custom.py
@@ -638,7 +638,8 @@ def create_vm(cmd, vm_name, resource_group_name, image=None, size='Standard_DS1_
         attach_os_disk=attach_os_disk, os_disk_size_gb=os_disk_size_gb, custom_data=custom_data, secrets=secrets,
         license_type=license_type, zone=zone, disk_info=disk_info,
         boot_diagnostics_storage_uri=boot_diagnostics_storage, ultra_ssd_enabled=ultra_ssd_enabled,
-        proximity_placement_group=proximity_placement_group, computer_name=computer_name, dedicated_host=dedicated_host)
+        proximity_placement_group=proximity_placement_group, computer_name=computer_name,
+        dedicated_host=dedicated_host, priority=priority, max_billing=max_billing, eviction_policy=eviction_policy)
 
     vm_resource['dependsOn'] = vm_dependencies